+++ src/Profiles/Models/Profile.php
@@ -2,18 +2,20 @@
 
 namespace Enraiged\Profiles\Models;
 
+use Enraiged\Avatars\Contracts\AvatarableContract;
+use Enraiged\Avatars\Traits\Avatarable;
 use Enraiged\Database\Tracking;
 use Enraiged\Profiles\Factories\ProfileFactory;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Database\Eloquent\Model;
 
-class Profile extends Model
+class Profile extends Model implements AvatarableContract
 {
     use Attributes\Initials,
         Attributes\Name,
         Relations\HasOneUser,
-        HasFactory, SoftDeletes, Tracking;
+        Avatarable, HasFactory, SoftDeletes, Tracking;
 
     /** @var  string  The database table name. */
     protected $table = 'profiles';